add fit pirls method
